+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * add :given DSL for dependent parameters
   * make `rescue_from` works for all maru router
   * support `with` option for `rescue_from` DSL
+  * bring maru's params parser to phoenix
 
 * Bugfix
   * allow set optional params to false and blank value by default